Attribute Macro leptos_macro::server

#[server]

Declares that a function is a server function. This means that its body will only run on the server, i.e., when the ssr feature is enabled.

If you call a server function from the client (i.e., when the csr or hydrate features are enabled), it will instead make a network request to the server.

You can specify one, two, three, or four arguments to the server function. All of these arguments are optional.

A type name that will be used to identify and register the server function (e.g., MyServerFn). Defaults to a PascalCased version of the function name.

A URL prefix at which the function will be mounted when it’s registered (e.g., "/api"). Defaults to "/api".

The encoding for the server function ("Url", "Cbor", "GetJson", or "GetCbor“. See Server Function Encodings below.)

A specific endpoint path to be used in the URL. (By default, a unique path will be generated.)

// will generate a server function at `/api-prefix/hello`
#[server(MyServerFnType, "/api-prefix", "Url", "hello")]
pub async fn my_server_fn_type() /* ... */

// will generate a server function with struct `HelloWorld` and path
// `/api/hello2349232342342` (hash based on location in source)
#[server]
pub async fn hello_world() /* ... */

The server function itself can take any number of arguments, each of which should be serializable and deserializable with serde.

#[server(ReadPosts, "/api")]
pub async fn read_posts(how_many: u8, query: String) -> Result<Vec<Post>, ServerFnError> {
  // do some work on the server to access the database
  todo!()
}

Note the following:

Server functions must be async. Even if the work being done inside the function body can run synchronously on the server, from the client’s perspective it involves an asynchronous function call.

Server functions must return Result<T, ServerFnError>. Even if the work being done inside the function body can’t fail, the processes of serialization/deserialization and the network call are fallible.

Return types must implement Serialize. This should be fairly obvious: we have to serialize arguments to send them to the server, and we need to deserialize the result to return it to the client.

Arguments must implement Serialize and DeserializeOwned. They are serialized as an application/x-www-form-urlencoded form data using serde_qs or as application/cbor using cbor. Note: You should explicitly include serde with the derive feature enabled in your Cargo.toml. You can do this by running cargo add serde --features=derive.

Context comes from the server. Server functions are provided access to the HTTP request and other relevant server data via the server integrations, but they do not have access to reactive state that exists in the client.

Your server must be ready to handle the server functions at the API prefix you list. The easiest way to do this is to use the handle_server_fns function from leptos_actix or leptos_axum.

Server functions must have unique paths. Unique paths are automatically generated for each server function. If you choose to specify a path in the fourth argument, you must ensure that these are unique. You cannot define two server functions with the same URL prefix and endpoint path, even if they have different URL encodings, e.g. a POST method at /api/foo and a GET method at /api/foo.

Server Function Encodings

By default, the server function call is a POST request that serializes the arguments as URL-encoded form data in the body of the request. But there are a few other methods supported. Optionally, we can provide another argument to the #[server] macro to specify an alternate encoding:

#[server(AddTodo, "/api", "Url")]
#[server(AddTodo, "/api", "GetJson")]
#[server(AddTodo, "/api", "Cbor")]
#[server(AddTodo, "/api", "GetCbor")]

The four options use different combinations of HTTP verbs and encoding methods:

Name	Method	Request	Response

Url (default)	POST	URL encoded	JSON

GetJson	GET	URL encoded	JSON

Cbor	POST	CBOR	CBOR

GetCbor	GET	URL encoded	CBOR

In other words, you have two choices:

GET or POST? This has implications for things like browser or CDN caching; while POST requests should not be cached, GET requests can be.

Plain text (arguments sent with URL/form encoding, results sent as JSON) or a binary format (CBOR, encoded as a base64 string)?

Why not PUT or DELETE? Why URL/form encoding, and not JSON?**

These are reasonable questions. Much of the web is built on REST API patterns that encourage the use of semantic HTTP methods like DELETE to delete an item from a database, and many devs are accustomed to sending data to APIs in the JSON format.

The reason we use POST or GET with URL-encoded data by default is the <form> support. For better or for worse, HTML forms don’t support PUT or DELETE, and they don’t support sending JSON. This means that if you use anything but a GET or POST request with URL-encoded data, it can only work once WASM has loaded.

The CBOR encoding is suported for historical reasons; an earlier version of server functions used a URL encoding that didn’t support nested objects like structs or vectors as server function arguments, which CBOR did. But note that the CBOR forms encounter the same issue as PUT, DELETE, or JSON: they do not degrade gracefully if the WASM version of your app is not available.